requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-WTF>=1.2
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1
reportlab>=4.0
email-validator>=2.0